refactor: shorten ts imports
